Move dependency configuration classes in Shared Kernel to Configuration namespace
