+++ application/back-office/Api/Program.cs
@@ -1,5 +1,5 @@
 using PlatformPlatform.BackOffice;
-using PlatformPlatform.SharedKernel;
+using PlatformPlatform.SharedKernel.Configuration;
 using PlatformPlatform.SharedKernel.SinglePageApp;
 
 var builder = WebApplication.CreateBuilder(args);
