+++ application/back-office/Workers/Program.cs
@@ -1,6 +1,6 @@
 using PlatformPlatform.BackOffice;
 using PlatformPlatform.BackOffice.Database;
-using PlatformPlatform.SharedKernel;
+using PlatformPlatform.SharedKernel.Configuration;
 using PlatformPlatform.SharedKernel.Database;
 
 // Worker service is using WebApplication.CreateBuilder instead of Host.CreateDefaultBuilder to allow scaling to zero
